Image path changed / Security bug fixed
urllib not only opens http:// or https:// URLs, but also ftp:// and file://. With this, it might be possible to open local files on the executing machine which might be a security risk if the URL to open can be manipulated by an external user. The urllib.request module defines functions and classes which help in opening URLs. urllib.request.open can open ftp:// and file:// URLs. This is usually not intended and makes the application vulnerable to Server Side Request Forgery attack. Performing requests from user-provided data could allow attackers to make requests on the internal network or change, retrieve or delete sensitive information. You are yourself responsible for validating the URL before opening it with urllib.